# src/module/user/me.controller.ts
import {
  Controller,
  UseGuards,
  Get,
  HttpCode,
  Put,
  Post,
  Body,
  Delete,
  Query,
  Param,
  ParseUUIDPipe,
} from '@nestjs/common';
import {
  ApiTags,
  ApiBearerAuth,
  ApiOperation,
  ApiResponse,
} from '@nestjs/swagger';
import { UserService } from './user.service';
import { CurrentUser } from 'src/core/decorator';
import { JwtGuard } from 'src/core/guard';
import { CursorRequest } from 'src/shared/request/cursor.request';
import { PageRequest } from 'src/shared/request/page.request';
import {
  UpdateUserRequest,
  UpdateProfileImageRequest,
  UpdateBackgroundImageRequest,
  UpdateSubscriptionRequest,
  GetFollowingRequest,
  RegisterPushTokenRequest,
  VerifyIdentityRequest,
  UpsertCommissionNoticeRequest,
} from './dto/user.request';
import { AlbumBaseResponse } from '../album/dto/album.base.response';
import {
  MyProfileResponse,
  UpdateProfileConflictResponse,
  SubscriptionResponse,
  MyFollowersResponse,
  MyFollowingsResponse,
  MyBlockingsResponse,
  MyIdentityVerificationResponse,
  CommissionNoticeResponse,
} from './dto/user.response';
import {
  VerifyIdentity409Response,
  VerifyIdentity422Response,
} from './dto/identity-verification.error';
import { MyLikeFeedsResponse } from '../feed/dto/feed.response';
import { MySavePostsResponse } from '../post/dto/post.response';

@ApiTags('me')
@ApiBearerAuth()
@ApiResponse({ status: 401, description: 'Unauthorized' })
@ApiResponse({ status: 400, description: '유효성 검사 실패' })
@UseGuards(JwtGuard)
@Controller('me')
export class MeController {
  constructor(private readonly userService: UserService) {}

  @ApiOperation({ summary: '내 정보 조회' })
  @ApiResponse({ status: 200, type: MyProfileResponse })
  @ApiResponse({ status: 404, description: '없는 유저' })
  @Get()
  async getMe(@CurrentUser() userId: string): Promise<MyProfileResponse> {
    return this.userService.getMyProfile(userId);
  }

  @ApiOperation({ summary: '내 정보 변경' })
  @ApiResponse({ status: 204 })
  @ApiResponse({ status: 409, type: UpdateProfileConflictResponse })
  @HttpCode(204)
  @Put()
  async updateProfile(
    @CurrentUser() userId: string,
    @Body() dto: UpdateUserRequest,
  ) {
    await this.userService.updateProfile(userId, dto);
    return;
  }

  @ApiOperation({ summary: '회원 탈퇴' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Delete()
  async deleteUser(@CurrentUser() userId: string) {
    await this.userService.deleteMe(userId);
    return;
  }

  @ApiOperation({ summary: '프로필 이미지 변경' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Put('image')
  async updateProfileImage(
    @CurrentUser() userId: string,
    @Body() { imageName }: UpdateProfileImageRequest,
  ) {
    await this.userService.updateProfileImage(userId, imageName);
    return;
  }

  @ApiOperation({ summary: '푸시토큰 등록' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Put('push-token')
  async registerPushToken(
    @CurrentUser() userId: string,
    @Body() dto: RegisterPushTokenRequest,
  ) {
    await this.userService.registerPushToken({
      userId,
      ...dto,
    });
    return;
  }

  @ApiOperation({ summary: '프로필 이미지 삭제' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Delete('image')
  async deleteProfileImage(@CurrentUser() userId: string) {
    await this.userService.updateProfileImage(userId, null);
    return;
  }

  @ApiOperation({ summary: '배경사진 변경' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Put('background')
  async updateBackgroundImage(
    @CurrentUser() userId: string,
    @Body() { imageName }: UpdateBackgroundImageRequest,
  ) {
    await this.userService.updateBackgroundImage(userId, imageName);
    return;
  }

  @ApiOperation({ summary: '배경사진 삭제' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Delete('background')
  async deleteBackgroundImage(@CurrentUser() userId: string) {
    await this.userService.updateBackgroundImage(userId, null);
    return;
  }

  @ApiOperation({ summary: '알림 구독 여부 조회' })
  @ApiResponse({ status: 200, type: SubscriptionResponse })
  @Get('subscribe')
  async getSubscriptions(
    @CurrentUser() userId: string,
  ): Promise<SubscriptionResponse> {
    return await this.userService.getSubscription(userId);
  }

  @ApiOperation({ summary: '알림 구독 여부 수정' })
  @ApiResponse({ status: 204 })
  @Put('subscribe')
  @HttpCode(204)
  async updateSubscriptions(
    @CurrentUser() userId: string,
    @Body() { subscription }: UpdateSubscriptionRequest,
  ) {
    await this.userService.updateSubscription(userId, [
      ...new Set(subscription),
    ]);
    return;
  }

  @ApiOperation({ summary: '내 팔로워 조회' })
  @ApiResponse({ status: 200, type: MyFollowersResponse })
  @Get('followers')
  async getMyFollowers(
    @CurrentUser() userId: string,
    @Query() query: CursorRequest,
  ): Promise<MyFollowersResponse> {
    return this.userService.getMyFollowers(userId, {
      cursor: query.cursor ?? null,
      size: query.size ?? 20,
    });
  }

  @ApiOperation({ summary: '내 팔로잉 조회' })
  @ApiResponse({ status: 200, type: MyFollowingsResponse })
  @Get('followings')
  async getMyFollowings(
    @CurrentUser() userId: string,
    @Query() query: GetFollowingRequest,
  ): Promise<MyFollowingsResponse> {
    return this.userService.getMyFollowings(userId, {
      keyword: query.keyword ?? null,
      cursor: query.cursor ?? null,
      size: query.size ?? 20,
    });
  }

  @ApiOperation({ summary: '내가 블락한 유저 조회' })
  @ApiResponse({ status: 200, type: MyBlockingsResponse })
  @Get('blockings')
  async getMyBlockings(
    @CurrentUser() userId: string,
  ): Promise<MyBlockingsResponse> {
    return await this.userService.getMyBlockings(userId);
  }

  @ApiOperation({ summary: '내 팔로워 삭제' })
  @ApiResponse({ status: 204 })
  @ApiResponse({ status: 404, description: '팔로우하지 않은 유저' })
  @HttpCode(204)
  @Delete('followers/:id')
  async deleteMyFollower(
    @CurrentUser() userId: string,
    @Param('id', ParseUUIDPipe) targetId: string,
  ) {
    await this.userService.unfollowTransaction(targetId, userId);
    return;
  }

  @ApiOperation({ summary: '내 좋아요한 피드 조회' })
  @ApiResponse({ status: 200, type: MyLikeFeedsResponse })
  @Get('like-feeds')
  async getMyLikeFeeds(
    @CurrentUser() userId: string,
    @Query() query: CursorRequest,
  ): Promise<MyLikeFeedsResponse> {
    return await this.userService.getMyLikeFeeds(userId, {
      cursor: query.cursor ?? null,
      size: query.size ?? 20,
    });
  }

  @ApiOperation({ summary: '내가 저장한 피드 조회' })
  @ApiResponse({ status: 200, type: MyLikeFeedsResponse })
  @Get('save-feeds')
  async getMySaveFeeds(
    @CurrentUser() userId: string,
    @Query() query: CursorRequest,
  ): Promise<MyLikeFeedsResponse> {
    return await this.userService.getMySaveFeeds(userId, {
      cursor: query.cursor ?? null,
      size: query.size ?? 20,
    });
  }

  @ApiOperation({ summary: '내가 저장한 게시글 조회' })
  @ApiResponse({ status: 200, type: MySavePostsResponse })
  @Get('save-posts')
  async getMySavePosts(
    @CurrentUser() userId: string,
    @Query() { page, size }: PageRequest,
  ): Promise<MySavePostsResponse> {
    return await this.userService.getMySavePosts({
      userId,
      page: page ?? 1,
      size: size ?? 10,
    });
  }

  @ApiOperation({ summary: '내 앨범 목록 조회' })
  @ApiResponse({ status: 200, type: [AlbumBaseResponse] })
  @Get('albums')
  async getMyAlbums(
    @CurrentUser() userId: string,
  ): Promise<AlbumBaseResponse[]> {
    return await this.userService.getAlbumsByUserId(userId);
  }

  @ApiOperation({
    summary: '본인인증 완료 처리',
    description:
      '프론트에서 포트원 SDK로 본인인증을 완료한 후 받은 identityVerificationId를 전달하면, 서버가 포트원 API로 검증/저장한다.',
  })
  @ApiResponse({ status: 204 })
  @ApiResponse({ status: 409, type: VerifyIdentity409Response })
  @ApiResponse({ status: 422, type: VerifyIdentity422Response })
  @HttpCode(204)
  @Post('identity-verification')
  async verifyIdentity(
    @CurrentUser() userId: string,
    @Body() { identityVerificationId }: VerifyIdentityRequest,
  ) {
    await this.userService.verifyIdentity(userId, identityVerificationId);
    return;
  }

  @ApiOperation({ summary: '내 본인인증 상태 조회' })
  @ApiResponse({ status: 200, type: MyIdentityVerificationResponse })
  @Get('identity-verification')
  async getMyIdentityVerification(
    @CurrentUser() userId: string,
  ): Promise<MyIdentityVerificationResponse> {
    return await this.userService.getMyIdentityVerification(userId);
  }

  @ApiOperation({ summary: '내 커미션 공지 조회' })
  @ApiResponse({ status: 200, type: CommissionNoticeResponse })
  @Get('commission-notice')
  async getMyCommissionNotice(
    @CurrentUser() userId: string,
  ): Promise<CommissionNoticeResponse> {
    return await this.userService.findCommissionNoticeByUserId(userId);
  }

  @ApiOperation({
    summary: '내 커미션 공지 등록/수정 (본인인증 필요)',
    description:
      '본인인증이 완료된 유저만 가능. 1인 1공지로 동일 유저가 다시 호출하면 덮어쓴다.',
  })
  @ApiResponse({ status: 200, type: CommissionNoticeResponse })
  @ApiResponse({
    status: 422,
    type: VerifyIdentity422Response,
    description: 'NOT_VERIFIED — 본인인증 필요',
  })
  @Put('commission-notice')
  async upsertMyCommissionNotice(
    @CurrentUser() userId: string,
    @Body() dto: UpsertCommissionNoticeRequest,
  ): Promise<CommissionNoticeResponse> {
    return await this.userService.upsertCommissionNotice(userId, dto);
  }

  @ApiOperation({ summary: '내 커미션 공지 삭제' })
  @ApiResponse({ status: 204 })
  @HttpCode(204)
  @Delete('commission-notice')
  async deleteMyCommissionNotice(@CurrentUser() userId: string) {
    await this.userService.deleteCommissionNotice(userId);
    return;
  }
}